feat: 도메인 전체 구조 및 파일 추가
