+++ src/main/java/com/example/eat_together/domain/user/controller/UserController.java
@@ -1,0 +1,4 @@
+package com.example.eat_together.domain.user.controller;
+
+public class UserController {
+}
