+++ src/main/java/com/example/eat_together/domain/user/dto/request/UserRequestDto.java
@@ -1,0 +1,4 @@
+package com.example.eat_together.domain.user.dto.request;
+
+public class UserRequestDto {
+}
