+++ src/main/java/com/example/eat_together/domain/user/dto/response/UserResponseDto.java
@@ -1,0 +1,4 @@
+package com.example.eat_together.domain.user.dto.response;
+
+public class UserResponseDto {
+}
