+++ src/main/java/com/example/eat_together/domain/user/repository/UserRepository.java
@@ -1,0 +1,4 @@
+package com.example.eat_together.domain.user.repository;
+
+public class UserRepository {
+}
