+++ src/main/java/com/example/eat_together/domain/user/service/UserService.java
@@ -1,0 +1,4 @@
+package com.example.eat_together.domain.user.service;
+
+public class UserService {
+}
