--- src/main/java/com/example/eat_together/user/controller/UserController.java
@@ -1,4 +1,0 @@
-package com.example.eat_together.user.controller;
-
-public class UserController {
-}
