--- src/main/java/com/example/eat_together/user/dto/request/UserRequestDto.java
@@ -1,4 +1,0 @@
-package com.example.eat_together.user.dto.request;
-
-public class UserRequestDto {
-}
