--- src/main/java/com/example/eat_together/user/dto/response/UserResponseDto.java
@@ -1,4 +1,0 @@
-package com.example.eat_together.user.dto.response;
-
-public class UserResponseDto {
-}
