--- src/main/java/com/example/eat_together/user/repository/UserRepository.java
@@ -1,4 +1,0 @@
-package com.example.eat_together.user.repository;
-
-public class UserRepository {
-}
