--- src/main/java/com/example/eat_together/user/service/UserService.java
@@ -1,4 +1,0 @@
-package com.example.eat_together.user.service;
-
-public class UserService {
-}
